requirement.txt
fastapi
uvicorn
pdfplumber
python-multipart
python-dotenv
loguru
requests
python-dotenv

numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0


nltk>=3.8.0
spacy>=3.6.0
vaderSentiment>=3.3.2
textstat>=0.7.3
transformers>=4.30.0
torch>=2.0.0

librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1
ffmpeg-python>=0.2.0

praat-parselmouth>=0.4.3
audioread>=3.0.0


opencv-python>=4.8.0
mediapipe>=0.10.0
face-recognition>=1.3.0


scikit-learn>=1.3.0
joblib>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0

shap>=0.42.0
lime>=0.2.0


matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
Pillow>=10.0.0
kaleido>=0.2.1  


tqdm>=4.65.0
python-dotenv>=1.0.0



pytest>=7.4.0
pytest-asyncio>=0.21.0